pyproject.toml
[project]
name = "resume-graph"
version = "0.1.0"
description = "HR Skill Graph Tool — Neo4j + LlamaIndex + FastAPI"
requires-python = ">=3.12"
dependencies = [
    # LlamaIndex
    "llama-index-core>=0.12",
    "llama-index-graph-stores-neo4j>=0.4",
    "llama-index-llms-azure-openai>=0.3",
    "llama-index-embeddings-azure-openai>=0.3",
    "llama-index-readers-file>=0.4",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    # MCP
    "mcp>=1.0",
    # Config & validation
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # PDF generation
    "reportlab>=4.0",
    # Neo4j driver (used by LlamaIndex, also direct access)
    "neo4j>=5.0",
    # LLM SDK (for profile generation)
    "openai>=1.0",
    # HTTP client (for MCP -> API calls)
    "httpx>=0.28",
    "pymupdf>=1.27.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resume_graph"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
